+++ src/Components/Element.php
@@ -18,6 +18,8 @@ use Termwind\ValueObjects\Styles;
  * @method Element italic()
  * @method Element underline()
  * @method Element lineThrough()
+ * @method int getLength()
+ * @method int getInnerWidth()
  * @method array getProperties()
  * @method Element href(string $href)
  * @method bool hasStyle(string $style)