+++ src/ValueObjects/Styles.php
@@ -25,7 +25,7 @@ final class Styles
     /**
      * Finds all the styling on a string.
      */
-    private const STYLING_REGEX = "/\<[\w=#\/\;,]+\>|\\e\[\d+m/";
+    public const STYLING_REGEX = "/\<[\w=#\/\;,:.&,%?]+\>|\\e\[\d+m/";
 
     /** @var array<int, string> */
     private array $styles = [];
@@ -121,27 +121,11 @@ final class Styles
     }
 
     /**
-     * Checks if formatter has styles.
-     */
-    final public function hasInheritableStyles(): bool
-    {
-        return ($this->properties['colors']['bg'] ?? []) !== []
-            || ($this->properties['colors']['fg'] ?? []) !== []
-            || ($this->properties['options'] ?? []) !== []
-            || ($this->properties['href'] ?? []) !== []
-            || ($this->properties['styles']['ml'] ?? 0) > 0
-            || ($this->properties['styles']['mr'] ?? 0) > 0
-            || ($this->properties['styles']['pl'] ?? 0) > 0
-            || ($this->properties['styles']['pr'] ?? 0) > 0
-            || ($this->properties['styles']['width'] ?? 0) > 0;
-    }
-
-    /**
      * Inherit styles from given Styles object.
      */
     final public function inheritFromStyles(Styles $styles): self
     {
-        foreach (['ml', 'mr', 'pl', 'pr', 'width'] as $style) {
+        foreach (['ml', 'mr', 'pl', 'pr', 'width', 'maxWidth', 'spaceY', 'spaceX'] as $style) {
             $this->properties['parentStyles'][$style] = array_merge(
                 $this->properties['parentStyles'][$style] ?? [],
                 $styles->properties['parentStyles'][$style] ?? []
@@ -149,6 +133,8 @@ final class Styles
 
             $this->properties['parentStyles'][$style][] = $styles->properties['styles'][$style] ?? 0;
         }
+
+        $this->properties['parentStyles']['justifyContent'] = $styles->properties['styles']['justifyContent'] ?? false;
 
         foreach (['bg', 'fg'] as $colorType) {
             $value = (array) ($this->properties['colors'][$colorType] ?? []);
@@ -324,7 +310,35 @@ final class Styles
      */
     final public function px(int $padding): self
     {
-        return $this->p($padding);
+        return $this->pl($padding)->pr($padding);
+    }
+
+    /**
+     * Adds the given padding top.
+     */
+    final public function pt(int $padding): static
+    {
+        return $this->with(['styles' => [
+            'pt' => $padding,
+        ]]);
+    }
+
+    /**
+     * Adds the given padding bottom.
+     */
+    final public function pb(int $padding): static
+    {
+        return $this->with(['styles' => [
+            'pb' => $padding,
+        ]]);
+    }
+
+    /**
+     * Adds the given vertical padding.
+     */
+    final public function py(int $padding): self
+    {
+        return $this->pt($padding)->pb($padding);
     }
 
     /**
@@ -332,7 +346,27 @@ final class Styles
      */
     final public function p(int $padding): self
     {
-        return $this->pl($padding)->pr($padding);
+        return $this->pt($padding)->pr($padding)->pb($padding)->pl($padding);
+    }
+
+    /**
+     * Adds the given vertical margin to the childs, ignoring the first child.
+     */
+    final public function spaceY(int $space): self
+    {
+        return $this->with(['styles' => [
+            'spaceY' => $space,
+        ]]);
+    }
+
+    /**
+     * Adds the given horizontal margin to the childs, ignoring the first child.
+     */
+    final public function spaceX(int $space): self
+    {
+        return $this->with(['styles' => [
+            'spaceX' => $space,
+        ]]);
     }
 
     /**
@@ -344,13 +378,12 @@ final class Styles
             throw new InvalidStyle('`border-t` can only be used on an "hr" element.');
         }
 
-        $this->styleModifiers[__METHOD__] = static function ($text, $styles): string {
-            $length = mb_strlen(preg_replace(self::STYLING_REGEX, '', $text), 'UTF-8');
-
+        $this->styleModifiers[__METHOD__] = function ($text, $styles): string {
+            $length = $this->getLength($text);
             if ($length < 1) {
                 $margins = (int) ($styles['ml'] ?? 0) + ($styles['mr'] ?? 0);
 
-                return str_repeat('─', terminal()->width() - $margins);
+                return str_repeat('─', self::getParentWidth($this->properties['parentStyles'] ?? []) - $margins);
             }
 
             return str_repeat('─', $length);
@@ -409,6 +442,16 @@ final class Styles
     final public function wFull(): static
     {
         return $this->w('1/1');
+    }
+
+    /**
+     * Defines a maximum width of an element.
+     */
+    final public function maxW(int|string $width): static
+    {
+        return $this->with(['styles' => [
+            'maxWidth' => $width,
+        ]]);
     }
 
     /**
@@ -475,6 +518,16 @@ final class Styles
     }
 
     /**
+     * Do not display element's content.
+     */
+    final public function hidden(): self
+    {
+        return $this->with(['styles' => [
+            'display' => 'hidden',
+        ]]);
+    }
+
+    /**
      * Makes a line break before the element's content.
      */
     final public function block(): self
@@ -482,6 +535,79 @@ final class Styles
         return $this->with(['styles' => [
             'display' => 'block',
         ]]);
+    }
+
+    /**
+     * Makes an element eligible to work with flex-1 element's style.
+     */
+    final public function flex(): self
+    {
+        return $this->with(['styles' => [
+            'display' => 'flex',
+        ]]);
+    }
+
+    /**
+     * Makes an element grow and shrink as needed, ignoring the initial size.
+     */
+    final public function flex1(): self
+    {
+        return $this->with(['styles' => [
+            'flex-1' => true,
+        ]]);
+    }
+
+    /**
+     * Justifies childs along the element with an equal amount of space between.
+     */
+    final public function justifyBetween(): self
+    {
+        return $this->with(['styles' => [
+            'justifyContent' => 'between',
+        ]]);
+    }
+
+    /**
+     * Justifies childs along the element with an equal amount of space between
+     * each item and half around.
+     */
+    final public function justifyAround(): self
+    {
+        return $this->with(['styles' => [
+            'justifyContent' => 'around',
+        ]]);
+    }
+
+    /**
+     * Justifies childs along the element with an equal amount of space around each item.
+     */
+    final public function justifyEvenly(): self
+    {
+        return $this->with(['styles' => [
+            'justifyContent' => 'evenly',
+        ]]);
+    }
+
+    /**
+     * Justifies childs along the center of the container’s main axis.
+     */
+    final public function justifyCenter(): self
+    {
+        return $this->with(['styles' => [
+            'justifyContent' => 'center',
+        ]]);
+    }
+
+    /**
+     * Repeats the string given until it fills all the content.
+     */
+    final public function contentRepeat(string $string): self
+    {
+        $string = preg_replace("/\[?'?([^'|\]]+)'?\]?/", '$1', $string) ?? '';
+
+        $this->textModifiers[__METHOD__] = static fn (): string => str_repeat($string, (int) floor(terminal()->width() / mb_strlen($string, 'UTF-8')));
+
+        return $this;
     }
 
     /**
@@ -545,6 +671,8 @@ final class Styles
      */
     final public function href(string $href): self
     {
+        $href = str_replace('%', '%%', $href);
+
         return $this->with(['href' => array_filter([$href])]);
     }
 
@@ -553,9 +681,6 @@ final class Styles
      */
     final public function format(string $content): string
     {
-        $display = $this->properties['styles']['display'] ?? 'inline';
-        $isFirstChild = (bool) $this->properties['isFirstChild'] ?? false;
-
         foreach ($this->textModifiers as $modifier) {
             $content = $modifier(
                 $content,
@@ -570,29 +695,7 @@ final class Styles
             $content = $modifier($content, $this->properties['styles'] ?? []);
         }
 
-        [$marginLeft, $marginRight] = $this->getMargins();
-        [$paddingLeft, $paddingRight] = $this->getPaddings();
-
-        $content = preg_replace('/\r[ \t]?/', "\n",
-            (string) preg_replace(
-                '/\n/',
-                str_repeat(' ', $marginRight + $paddingRight)
-                ."\n".
-                str_repeat(' ', $marginLeft + $paddingLeft),
-            $content)
-        );
-
-        return sprintf(
-            $this->getFormatString(),
-            $display === 'block' && ! $isFirstChild ? "\n" : '',
-            str_repeat("\n", (int) ($this->properties['styles']['mt'] ?? 0)),
-            str_repeat(' ', $marginLeft),
-            str_repeat(' ', $paddingLeft),
-            $content,
-            str_repeat(' ', $paddingRight),
-            str_repeat(' ', $marginRight),
-            str_repeat("\n", (int) ($this->properties['styles']['mb'] ?? 0)),
-        );
+        return $this->applyStyling($content);
     }
 
     /**
@@ -631,35 +734,47 @@ final class Styles
 
         // If there are no styles we don't need extra tags
         if ($styles === []) {
-            return '%s%s%s%s%s%s%s%s';
+            return '%s%s%s%s%s';
         }
 
-        return '%s%s%s<'.implode(';', $styles).'>%s%s%s</>%s%s';
+        return '%s<'.implode(';', $styles).'>%s%s%s</>%s';
     }
 
     /**
      * Get the margins applied to the element.
      *
-     * @return array{0: int, 1: int}
+     * @return array{0: int, 1: int, 2: int, 3: int}
      */
     private function getMargins(): array
     {
+        $isFirstChild = (bool) $this->properties['isFirstChild'];
+
+        $spaceY = $this->properties['parentStyles']['spaceY'] ?? [];
+        $spaceY = ! $isFirstChild ? end($spaceY) : 0;
+
+        $spaceX = $this->properties['parentStyles']['spaceX'] ?? [];
+        $spaceX = ! $isFirstChild ? end($spaceX) : 0;
+
         return [
-            $this->properties['styles']['ml'] ?? 0,
+            $spaceY > 0 ? $spaceY : $this->properties['styles']['mt'] ?? 0,
             $this->properties['styles']['mr'] ?? 0,
+            $this->properties['styles']['mb'] ?? 0,
+            $spaceX > 0 ? $spaceX : $this->properties['styles']['ml'] ?? 0,
         ];
     }
 
     /**
      * Get the paddings applied to the element.
      *
-     * @return array{0: int, 1: int}
+     * @return array{0: int, 1: int, 2: int, 3: int}
      */
     private function getPaddings(): array
     {
         return [
-            $this->properties['styles']['pl'] ?? 0,
+            $this->properties['styles']['pt'] ?? 0,
             $this->properties['styles']['pr'] ?? 0,
+            $this->properties['styles']['pb'] ?? 0,
+            $this->properties['styles']['pl'] ?? 0,
         ];
     }
 
@@ -669,10 +784,15 @@ final class Styles
     private function applyWidth(string $content): string
     {
         $styles = $this->properties['styles'] ?? [];
-        $width = $styles['width'] ?? 0;
+        $width = $styles['width'] ?? -1;
+        $maxWidth = $styles['maxWidth'] ?? 0;
 
-        if ($width < 1) {
+        if ($width < 0) {
             return $content;
+        }
+
+        if ($width === 0) {
+            return '';
         }
 
         if (is_string($width)) {
@@ -683,12 +803,17 @@ final class Styles
             );
         }
 
+        if ($maxWidth > 0) {
+            $width = min($styles['maxWidth'], $width);
+        }
+
         $width -= ($styles['pl'] ?? 0) + ($styles['pr'] ?? 0);
-        $length = mb_strlen(preg_replace(self::STYLING_REGEX, '', $content) ?? '', 'UTF-8');
+        $length = $this->getLength($content);
 
         preg_match_all("/\n+/", $content, $matches);
+
         $width *= count($matches[0] ?? []) + 1;
-        $width += count($matches[0] ?? []);
+        $width += mb_strlen($matches[0][0] ?? '', 'UTF-8');
 
         if ($length <= $width) {
             $space = $width - $length;
@@ -701,6 +826,96 @@ final class Styles
         }
 
         return self::trimText($content, $width);
+    }
+
+    /**
+     * It applies the styling for the content.
+     */
+    private function applyStyling(string $content): string
+    {
+        $display = $this->properties['styles']['display'] ?? 'inline';
+
+        if ($display === 'hidden') {
+            return '';
+        }
+
+        $isFirstChild = (bool) $this->properties['isFirstChild'];
+
+        [$marginTop, $marginRight, $marginBottom, $marginLeft] = $this->getMargins();
+        [$paddingTop, $paddingRight, $paddingBottom, $paddingLeft] = $this->getPaddings();
+
+        $content = (string) preg_replace('/\r[ \t]?/', "\n",
+            (string) preg_replace(
+                '/\n/',
+                str_repeat(' ', $marginRight + $paddingRight)
+                ."\n".
+                str_repeat(' ', $marginLeft + $paddingLeft),
+            $content)
+        );
+
+        $formatted = sprintf(
+            $this->getFormatString(),
+            str_repeat(' ', $marginLeft),
+            str_repeat(' ', $paddingLeft),
+            $content,
+            str_repeat(' ', $paddingRight),
+            str_repeat(' ', $marginRight),
+        );
+
+        $empty = str_replace(
+            $content,
+            str_repeat(' ', $this->getLength($content)),
+            $formatted
+        );
+
+        $items = [];
+
+        if (in_array($display, ['block', 'flex'], true) && ! $isFirstChild) {
+            $items[] = "\n";
+        }
+
+        if ($marginTop > 0) {
+            $items[] = str_repeat("\n", $marginTop);
+        }
+
+        if ($paddingTop > 0) {
+            $items[] = $empty."\n";
+        }
+
+        $items[] = $formatted;
+
+        if ($paddingBottom > 0) {
+            $items[] = "\n".$empty;
+        }
+
+        if ($marginBottom > 0) {
+            $items[] = str_repeat("\n", $marginBottom);
+        }
+
+        return implode('', $items);
+    }
+
+    /**
+     * Get the length of the text provided without the styling tags.
+     */
+    public function getLength(string $text = null): int
+    {
+        return mb_strlen(preg_replace(
+            self::STYLING_REGEX,
+            '',
+            $text ?? $this->element?->toString() ?? ''
+        ) ?? '', 'UTF-8');
+    }
+
+    /**
+     * Get the length of the element without margins.
+     */
+    public function getInnerWidth(): int
+    {
+        $innerLength = $this->getLength();
+        [, $marginRight, , $marginLeft] = $this->getMargins();
+
+        return $innerLength - $marginLeft - $marginRight;
     }
 
     /**
@@ -752,11 +967,11 @@ final class Styles
      *
      * @param  array<string, array<int|string>>  $styles
      */
-    private static function getParentWidth(array $styles): int
+    public static function getParentWidth(array $styles): int
     {
         $width = terminal()->width();
-
         foreach ($styles['width'] ?? [] as $index => $parentWidth) {
+            $maxWidth = (int) $styles['maxWidth'][$index];
             $margins = (int) $styles['ml'][$index] + (int) $styles['mr'][$index];
 
             if ($parentWidth < 1) {
@@ -770,6 +985,10 @@ final class Styles
             $width = count($matches) !== 3
                 ? (int) $parentWidth
                 : (int) floor($width * $matches[1] / $matches[2]);
+
+            if ($maxWidth > 0) {
+                $width = min($maxWidth, $width);
+            }
 
             $width -= $margins;
             $width -= (int) $styles['pl'][$index] + (int) $styles['pr'][$index];
@@ -788,7 +1007,7 @@ final class Styles
         $text = rtrim(mb_strimwidth(preg_replace(self::STYLING_REGEX, '', $text) ?? '', 0, $width, '', 'UTF-8'));
 
         foreach ($matches[0] ?? [] as [$part, $index]) {
-            $text = mb_substr($text, 0, $index, 'UTF-8').$part.mb_substr($text, $index, null, 'UTF-8');
+            $text = substr($text, 0, $index).$part.substr($text, $index, null);
         }
 
         return $text;